% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcShift.R
\name{calcShift}
\alias{calcShift}
\title{Create a decile plot and a scatter for CNV versus flanking region in an offspring. Uses rogme package.}
\usage{
calcShift(lrr_dt)
}
\arguments{
\item{lrr_dt}{A data table with probe level LRR values for all individuals in a trio.}
}
\value{
A list with a data table with sampling results and a graph object
}
\description{
Create a decile plot and a scatter for CNV versus flanking region in an offspring. Uses rogme package.
}
