% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifySingles.R
\name{classifySingles}
\alias{classifySingles}
\title{Main function that uses prepared summary statistics to classify CNVs into categories}
\usage{
classifySingles(analyzed_df, lrr_flank_pass = 0.3)
}
\arguments{
\item{analyzed_df}{Output of runAnalyzeSignal for a single sample mode}

\item{lrr_flank_pass}{A cutoff on the LRR_SD in flanking region, default 0.3 in single mode}
}
\value{
A data table with most useful statistics and classification results per individual per CNV
}
\description{
Main function that uses prepared summary statistics to classify CNVs into categories
}
