% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyTrios.R
\name{classifyTrios}
\alias{classifyTrios}
\title{Main function that uses prepared summary statistics to classify CNVs into categories}
\usage{
classifyTrios(analyzed_df, lrr_flank_pass = 0.2)
}
\arguments{
\item{analyzed_df}{Output of runAnalyzeSignal}

\item{lrr_flank_pass}{A cutoff on the LRR_SD in flanking region that defines placement of putative de novo CNVs into Unlikely category for more detailed examination}
}
\value{
A data table with most useful statistics per individual in a trio and classification results for an offspring CNV
}
\description{
Main function that uses prepared summary statistics to classify CNVs into categories
}
