% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterBAF.R
\name{clusterBAF}
\alias{clusterBAF}
\title{Cluster BAF values using given cluster centers, helper function}
\usage{
clusterBAF(mybaf, margin.dup = 0.05, margin.het = 0.1, margin.nh = 0.17)
}
\arguments{
\item{mybaf}{The input data table collated with readSiftInput for individual CNV and sample in "data".}

\item{margin.dup}{A margin value around dup clusters AAB or ABB (0.33 and 0.66). Default is 0.05.}

\item{margin.het}{A margin value around heterozygous AB cluster (0.5). Default is 0.1.}

\item{margin.nh}{A margin value around homozygous AA or BB clusters (0 and 1). Default is 0.17}
}
\value{
A data table with counts and fractions of probes in each cluster.
}
\description{
Cluster BAF values using given cluster centers, helper function
}
