% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decodeHMMstates.R
\name{decodeHMMstates}
\alias{decodeHMMstates}
\title{Create a table mapping PennCNV HMM states to copy numbers}
\usage{
decodeHMMstates()
}
\value{
A data.frame with the mapping of HMM states in PennCNV to copynumber states
}
\description{
Create a table mapping PennCNV HMM states to copy numbers
}
\examples{
decodeHMMstates()
}
