% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractInheritance.R
\name{extractInheritance}
\alias{extractInheritance}
\title{Annotate original PennCNV trio data table with inheritance status using trio state decomposition}
\usage{
extractInheritance(penntriotable)
}
\arguments{
\item{penntriotable}{A data.frame read into R with readPennTrioOriginal}
}
\value{
A data.frame with status of inheritance, one of "denovo"/"inherited"/"ambiguous" for offspring and "call_in_parent" for parents
}
\description{
Annotate original PennCNV trio data table with inheritance status using trio state decomposition
}
