% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePythonCommands.R
\name{makePythonCommands}
\alias{makePythonCommands}
\title{Create commands to run helper python scripts that automate SNP extraction via PennCNV infer_snp_allele.pl for a cohort.}
\usage{
makePythonCommands(
  penn_path,
  pfb_file,
  penn_trio_list,
  triocnv_file,
  n_flanking_snp = 60,
  dataset = "TEST",
  run_dir
)
}
\arguments{
\item{penn_path}{Full path to PennCNV installation folder}

\item{pfb_file}{Full path to PFB file used in the analysis}

\item{penn_trio_list}{Full path to tab-separated file, listing one trio per line in order father,mother,offspring with full paths to raw data files for each, as used in PennCNV trio}

\item{triocnv_file}{Full path to annotated trio file for offspring (after inheritance map)}

\item{n_flanking_snp}{Integer. A number of probes to extract in upstream and downstream flanks of CNV. Default is 60.}

\item{dataset}{Dataset string used throughout the analysis. Must be consistent. Default is TEST.}

\item{run_dir}{Full path to where extract the snp data. Will be incorporated into extract commands.}
}
\value{
A data table with merged sample and CNV ids as well as original tristates and counts of segments merged
}
\description{
Create commands to run helper python scripts that automate SNP extraction via PennCNV infer_snp_allele.pl for a cohort.
}
