% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapTriostate.R
\name{mapTriostate}
\alias{mapTriostate}
\title{Link inheritance status and original trio state as well as merging tracing}
\usage{
mapTriostate(orig_dt, merged_dt)
}
\arguments{
\item{orig_dt}{A table with output file of PennCNV trio}

\item{merged_dt}{A Table with the result of merging with PennCNV of the original file}
}
\value{
A data table with merged sample and CNV ids as well as original tristates and counts of segments merged
}
\description{
Link inheritance status and original trio state as well as merging tracing
}
