% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCohort.R
\name{plotCohort}
\alias{plotCohort}
\title{Write the pdf files with the visualisations of local CNV SNP data for each trio in the data}
\usage{
plotCohort(
  main_data,
  sifted_data,
  classified_data,
  output_dir,
  dataset,
  mono_marker = T,
  subset_nprobes = NULL,
  subset_length = NULL,
  single = FALSE
)
}
\arguments{
\item{main_data}{A list holding input data with one obligatory table "data" and two optional "qcsum" and "merge"
output of readInputs}

\item{sifted_data}{A table with summary statistis per individual in trio, output of runAnalyzeSignal}

\item{classified_data}{A table with final classifications, offspring-centered. Output of classifyTrios}

\item{output_dir}{Folder in which to write the plot files}

\item{dataset}{Dataset string to use in file naming}

\item{mono_marker}{(TRUE/FALSE) If the array contains non-polymorphic markers with BAF>1, do not plot these. TRUE by default}

\item{subset_nprobes}{Only plot the CNVs containing that many or more probes}

\item{subset_length}{Only plot the CNVs of the given legth (in bp) or longer}

\item{single}{Plot cohort of independent single samples}
}
\value{

}
\description{
Write the pdf files with the visualisations of local CNV SNP data for each trio in the data
}
