% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRawTrio.R
\name{plotRawTrio}
\alias{plotRawTrio}
\title{Visualize local CNV SNP data and key summary statistics for a trio}
\usage{
plotRawTrio(
  input_data,
  sifted_data,
  print_title = TRUE,
  penn_qcsum = NULL,
  mono_marker = TRUE,
  merge_trace = NULL
)
}
\arguments{
\item{input_data}{Data slot of an object read with readInputs, for a single CNV and individual}

\item{sifted_data}{Output of runAnalyzeSignal for the input data}

\item{print_title}{Print detailed summary statistics in a plot header. Default is TRUE.}

\item{penn_qcsum}{Table with PennCNV QC summaries stored in "qcsum" by readInputs if supplied. Will add global LRR_SD and WF values if present. Default is NULL.}

\item{mono_marker}{(TRUE/FALSE) If the array contains non-polymorphic markers with BAF>1, do not plot these. TRUE by default.}

\item{merge_trace}{Table in a custom format with original PennCNV trio status and merging statistics.}
}
\value{
A ggplot2 plot object
}
\description{
Visualize local CNV SNP data and key summary statistics for a trio
}
