% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareClassSingle.R
\name{prepareClassSingle}
\alias{prepareClassSingle}
\title{Prepare data summaries for final classifications, a helper function.
Performs main steps to check support of CNV in a single sample data}
\usage{
prepareClassSingle(analyzed_df)
}
\arguments{
\item{analyzed_df}{A dataframe processed with runAnalyseSignal in single sample mode}
}
\value{
A data.table formatted for final classification with classifySingles
}
\description{
Prepare data summaries for final classifications, a helper function.
Performs main steps to check support of CNV in a single sample data
}
