% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readInputs.R
\name{readInputs}
\alias{readInputs}
\title{Load and combine multiple input files}
\usage{
readInputs(args)
}
\arguments{
\item{args}{A list holding all necessary data files. eg  "triocnv_file" (input formatted penn triocnv file for offspring),
"probecoord_file" (file created by _runcommands script in flanks ),
"snp_flank_file" (_candidateCnv.txt file in flanks),
"snp_cnv_log_file" (_candidateCnvLOG.txt file in cnv locus),
"cnv_qcsum_file" (*.qcsum file from PennCNV), "dataset" (dataset string used throughout the analysis),
"cache_id" (full path to where to store rds cache files for computation havier steps),
"merge_trace" (table in a custom format with original PennCNV trio status and merging statistics, use runExtractInheritance to create this table)}
}
\value{
A list with "data": a data table with SNP level values of LRR and BAF for each individual in a trio;
if "cnv_qcsum_file" in args is supplied, a "qcsum" table with the genome-wide LRR and BAF summaries for each individual;
if "merge_trace" in args is supplied, a "merge" table with original PennCNV trio states and merging segment counts
}
\description{
Load and combine multiple input files
}
