% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPennTrioFormat.R
\name{readPennTrioFormat}
\alias{readPennTrioFormat}
\title{Load PennCNV trio adapted file into R. Expects the coding "sample":"relation":"status" in the fifth column.}
\usage{
readPennTrioFormat(filename)
}
\arguments{
\item{filename}{A path to the output file of PennCNV trio, produced in the infer inheritance steps.}
}
\value{
A table from the input  file
}
\description{
Load PennCNV trio adapted file into R. Expects the coding "sample":"relation":"status" in the fifth column.
}
