% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPennTrioOriginal.R
\name{readPennTrioOriginal}
\alias{readPennTrioOriginal}
\title{Load PennCNV trio file in original format as supplied by PennCNV into R}
\usage{
readPennTrioOriginal(filename, triostate = TRUE)
}
\arguments{
\item{filename}{A path to the output file of PennCNV trio}

\item{triostate}{Is a column with trio state present? Default TRUE.}
}
\value{
A table with the file content.
}
\description{
Load PennCNV trio file in original format as supplied by PennCNV into R
}
