% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSingle.R
\name{readSingle}
\alias{readSingle}
\title{Load and format data for single sample processing}
\usage{
readSingle(snp_file)
}
\arguments{
\item{snp_file}{A file with extracted snp data, after running 'extract_snp_single.py'}
}
\value{
A data.frame suitable for downstream single sample analysis
}
\description{
Load and format data for single sample processing
}
