% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stabilizeDistr.R
\name{stabilizeDistr}
\alias{stabilizeDistr}
\title{Remove outliers in the CNV locus LRR values, a helper function.}
\usage{
stabilizeDistr(ds)
}
\arguments{
\item{ds}{A vector of LRR values}
}
\value{
A vector of regularized LRR values
}
\description{
Remove outliers in the CNV locus LRR values, a helper function.
}
