% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statusFromTriostate.R
\name{statusFromTriostate}
\alias{statusFromTriostate}
\title{Extract inheritance status from penn trio string collapsing copy number strings to discrete "del", "dup", "normal" or ""copy-neutrLOH" following PennCNV encoding (Wang et al, 2007)}
\usage{
statusFromTriostate(penntriostate, offspring_order = 1)
}
\arguments{
\item{penntriostate}{A string of three digits representing copynumber state for members of a trio/quad}

\item{offspring_order}{Can be used both on trios and quads, therefore the index of an offspring (1 or 2) needs to be specified. Defaut is 1. (a trio)}
}
\value{
A string with a status, one of "denovo"/"inherited"
}
\description{
Extract inheritance status from penn trio string collapsing copy number strings to discrete "del", "dup", "normal" or ""copy-neutrLOH" following PennCNV encoding (Wang et al, 2007)
}
\examples{
statusFromTriostate("323")
}
