% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testLRRDistributions.R
\name{testLRRDistributions}
\alias{testLRRDistributions}
\title{Apply a row of statistical tests to LRR distribtuions in CNV and flanking regions within and between individuals in a trio}
\usage{
testLRRDistributions(my_lrr, cohen_dist = 3, sign_lrr_dist = 0.01)
}
\arguments{
\item{my_lrr}{The input data table with LRR values in CNV and flanking regions collated with readInputs in "data" for each individual}

\item{cohen_dist}{A cutoff on Cohen distance between CNV and flank LRR value distributions. If cohen distance is equal or smaller that this cutoff the distributions are considered not different without further testing. Default is 3.}

\item{sign_lrr_dist}{A cutoff of p-value at which difference between two individual's CNV LRR value distributions is set as significant, using Kernel density based global two-sample comparison test from ks package. Default is 0.01.}
}
\value{
A data table with summaries and test results for LRR-based tests between the individuals (lrrtest) and between CNV and flank for each individual (flanktest).
}
\description{
Apply a row of statistical tests to LRR distribtuions in CNV and flanking regions within and between individuals in a trio
}
