% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testLRRDistributionsSingle.R
\name{testLRRDistributionsSingle}
\alias{testLRRDistributionsSingle}
\title{Apply a row of statistical tests to LRR distribuions in CNV and flanking regions within a single individual}
\usage{
testLRRDistributionsSingle(my_lrr, cohen_dist = 3)
}
\arguments{
\item{my_lrr}{The input data table with LRR values in CNV and flanking regions for each individual}

\item{cohen_dist}{A cutoff on Cohen distance between CNV and flank LRR value distributions. If cohen distance is equal or smaller that this cutoff the distributions are considered not different without further testing. Default is 3.}
}
\value{
A data table with summaries and test results for LRR-based tests between CNV and flank for each individual (flanktest).
}
\description{
Apply a row of statistical tests to LRR distribuions in CNV and flanking regions within a single individual
}
