% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeSeecite.R
\name{writeSeecite}
\alias{writeSeecite}
\title{Write the CNV files with SeeCiTe assessment and all summary metrics in txt, plink and bed formats}
\usage{
writeSeecite(classified_data, output_dir, dataset)
}
\arguments{
\item{classified_data}{classified_data A table with final classifications, offspring-centered. Output of classifyTrios.}

\item{output_dir}{output_dir Folder in which to write the files.}

\item{dataset}{dataset Dataset string to use in file naming.}
}
\value{
No value returned. Writes files into a given folder.
}
\description{
Write the CNV files with SeeCiTe assessment and all summary metrics in txt, plink and bed formats
}
